requirements.txt
# ========================================
# Enterprise AI Resume Screening Platform
# Requirements File - Python 3.11 Compatible
# ========================================

# Core Data Science & ML
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0

# Machine Learning Models
xgboost==2.0.3
lightgbm==4.3.0

# Deep Learning & NLP
torch==2.2.0
spacy==3.7.4
sentence-transformers==2.3.1
transformers==4.37.2

# Explainable AI
shap==0.44.1
lime==0.2.0.1

# Fairness & Bias Detection
fairlearn==0.10.0

# MLOps & Experiment Tracking
mlflow==2.10.2
dvc==3.48.4

# API Development
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bleach==6.1.0

# Frontend Dashboard
streamlit==1.31.1

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0
openpyxl==3.1.2

# NLP & Text Processing
nltk==3.8.1

# Utilities
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1
joblib==1.3.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Production Server
gunicorn==21.2.0

# Logging & Monitoring
loguru==0.7.2

# Date & Time
python-dateutil==2.8.2

# HTML/Text Processing
beautifulsoup4==4.12.3
lxml==5.1.0

# Progress Bars
rich==13.7.0

flashtext>=2.7

